pyproject.toml
[project]
name = "telegram-for-content-download"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.24.0",
    "moviepy>=2.2.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "yt-dlp>=2026.2.4",
]
